src/fort/options/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# All the following settings are optional:
where = ["."]  # ["."] by default
include = ["fort_charis_options*"]  # ["*"] by default
exclude = ["tests*"]  # empty by default
namespaces = true  # true by default

[project]
name = "fort-charis-options"
version = "0.0.1"
authors = [
  { name="Alexandre Tuleu", email="alexandre.tuleu.2005@polytechnique.org" },
]
description = "C++ Options generator"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "libclang",
	"ccsyspath"
]
license = {file = "LICENSE"}

[project.scripts]
fort-charis-options = "fort_charis_options:main"

[project.urls]
Homepage = "https://github.com/formicidae-tracker/charis/src/fort/options"
Issues = "https://github.com/formicidae-tracker/charis/"